Adjusting tolerances looking at intgemm tests, bugfix on MSE

+++ src/impl_ruy-export.cpp
@@ -14,7 +14,7 @@ void quantize(const float *input, float scale, float zero_point, Index rows,
   // Dumb quantize we will improve this eventually.
   const Index size = rows * width;
   for (size_t i = 0; i < size; i++) {
-    float value = roundf(scale * input[i]);
+    float value = round(scale * input[i]);
     // int8 can't store larger than 127.0f.
     value = std::max(-127.0f, value);
     value = std::min(127.0f, value);
@@ -128,8 +128,8 @@ void int8MultiplyAndAddBias(const int8_t *input_A_prepared, float scale_A,
   for (size_t i = 0; i < rows_A; i++) {
     for (size_t j = 0; j < cols_B; j++) {
       Index idx = i * cols_B + j;
-      output[idx] = (static_cast<float>(dest_ptr[idx]) * unquant_multiplier) +
-                    input_bias_prepared[j];
+      output[idx] =
+          (dest_ptr[idx] * unquant_multiplier) + input_bias_prepared[j];
     }
   }
 }
